generalize crate for multi-device PCIe passthrough
